backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
alembic==1.13.1

# GIS Processing (use pip install wheel files on Windows)
# gdal - install from https://www.lfd.uci.edu/~gohlke/pythonlibs/
geopandas==0.14.2
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1

# Matching Engine
rapidfuzz==3.6.1

# File Processing
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# Report Generation
reportlab==4.0.8
jinja2==3.1.3

# Caching
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
orjson==3.9.12
